+++ automerge-backend/src/actor_map.rs
@@ -1,6 +1,5 @@
-use crate::protocol::{ElementID, Key};
 use std::cmp::Ordering;
-use automerge_protocol::{ActorID, OpID, ObjectID};
+use automerge_protocol::{ActorID, OpID, ObjectID, ElementID, Key};
 
 #[derive(PartialEq, Debug, Clone)]
 pub (crate) struct ActorMap(Vec<ActorID>);
