+++ automerge-backend/src/object_store.rs
@@ -3,9 +3,8 @@ use crate::error::AutomergeError;
 use crate::op_handle::OpHandle;
 use crate::ordered_set::{OrderedSet, SkipList};
 use crate::actor_map::ActorMap;
-use crate::protocol::{ElementID, Key};
 use im_rc::{HashMap, HashSet};
-use automerge_protocol::{ObjType, OpID};
+use automerge_protocol::{ObjType, OpID, ElementID, Key};
 
 /// ObjectHistory is what the OpSet uses to store operations for a particular
 /// key, they represent the two possible container types in automerge, a map or
@@ -45,7 +44,7 @@ impl ObjState {
     }
 
     fn get_parent(&self, id: &ElementID) -> Option<ElementID> {
-        self.insertions.get(&id).and_then(|i| i.key.as_element_id().ok())
+        self.insertions.get(&id).and_then(|i| i.key.as_element_id())
     }
 
     fn insertions_after(&self, parent: &ElementID) -> Vec<ElementID> {
