+++ automerge-protocol/src/serde_impls/key.rs
@@ -1,0 +1,18 @@
+use crate::{Key, ElementID};
+use serde::{Deserializer, Deserialize};
+use std::str::FromStr;
+
+impl<'de> Deserialize<'de> for Key {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let s = String::deserialize(deserializer)?;
+        if let Ok(eid) = ElementID::from_str(&s) {
+            Ok(Key::Seq(eid))
+        } else {
+            Ok(Key::Map(s))
+        }
+    }
+}
+
