+++ automerge-protocol/src/serde_impls/mod.rs
@@ -1,3 +1,5 @@
 mod change_hash;
 mod opid;
 mod object_id;
+mod element_id;
+mod key;
