+++ automerge-protocol/src/utility_impls/mod.rs
@@ -1,2 +1,4 @@
 mod opid;
 mod object_id;
+mod key;
+mod element_id;
